backend/requirements.txt
# Updated requirements with production dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart==0.0.20
sqlalchemy>=2.0.25
email-validator==2.3.0

# PostgreSQL support
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1

# Rate limiting
slowapi==0.1.9

# Logging
python-json-logger==3.2.1

# Environment variables
python-dotenv==1.2.1

# CORS
